Add access modifier exercises

+++ src/main/java/Exercises/ControlFlow/Number8/Main.java
@@ -3,7 +3,7 @@ package Exercises.ControlFlow.Number8;
 public class Main {
     public static void main(String[] args) {
         int numbers[] = { 1, 4, 3, 6, 8, 9};
-        double sum = 0;
+        int sum = 0;
         int count = 0;
 
         for (int number: numbers) {
@@ -14,7 +14,7 @@ public class Main {
         }
         double average = sum;
         if (count != 0) {
-            average = sum / count;
+            average = (double) sum / count;
         }
         System.out.println(average);
     }
